wrong type

+++ src/main.py
@@ -33,7 +33,7 @@ class TankDataTransform(BaseModel):
     InchesToESD: Optional[float]
     TimeUntilESD: Optional[float]
     Capacity: Optional[float]
-    PercentFull: Optional[float]
+    PercentFull: Optional[int]
 
 
 class TankType(Enum):
